Testing for the correct error value. These tests assume that symmetric heap has size < 1204000000 words.

+++ feature_tests/Fortran/alloc/test_shmem_shpalloc_03_char.f90
@@ -42,7 +42,7 @@ program test_shmem_shpalloc
   include 'shmem.fh'
 
 
-  integer*4, parameter :: nelems = 1024000000
+  integer, parameter :: nelems = 1024000000
 
   integer*8          :: array_addr
   character           :: array(1)
